Created SightingsForm component and linked to AddSighting container
