+++ client/src/Components/SightingsForm/SightingsForm.jsx
@@ -1,0 +1,175 @@
+import React, { useState } from "react";
+
+const SightingsForm = ({ handleFormSubmit }) => {
+  const [username, setUsername] = useState("");
+  const [title, setTitle] = useState("");
+  const [address, setAddress] = useState("");
+  const [city, setCity] = useState("");
+  const [state, setState] = useState("");
+  const [description, setDescription] = useState("");
+  const [imageUrl, setimageUrl] = useState("");
+
+  const showWidget = () => {
+    let widget = window.cloudinary.createUploadWidget(
+      {
+        cloudName: "dfqdc8mjr",
+        uploadPreset: "pmekgbnz",
+      },
+      (error, result) => {
+        if (!error && result && result.event === "success") {
+          console.log(result.info.url);
+          setimageUrl(result.info.url);
+        }
+      }
+    );
+    widget.open();
+  };
+
+  return (
+    <div>
+      <div className="columns">
+        <div className="column is-3 is-offset-3">
+          <div className="field">
+            <div id="photo-container">
+              <button onClick={showWidget}>Upload Photo</button>
+            </div>
+
+            <p className="control has-icons-left">
+              <input
+                id="imageUrl"
+                className="input"
+                type="text"
+                value={imageUrl}
+                onChange={(e) => setimageUrl(e.target.value)}
+              />
+              <span className="icon is-small is-left">
+                <i className="fa fa-camera"></i>
+              </span>
+            </p>
+          </div>
+          <span className="icon-text">
+            <span className="icon">
+              <i className="fa fa-camera"></i>
+            </span>
+            <span>Upload image</span>
+          </span>
+        </div>
+      </div>
+      <div className="columns">
+        <div className="column is-6 is-offset-3">
+          <div className="field">
+            <label className="label">Username</label>
+            <div className="control">
+              <input
+                id="username"
+                className="input"
+                type="text"
+                placeholder="Username"
+                value={username}
+                onChange={(e) => setUsername(e.target.value)}
+              />
+            </div>
+          </div>
+
+          <div className="field">
+            <label className="label">Title</label>
+            <div className="control">
+              <input
+                id="title"
+                className="input"
+                type="text"
+                placeholder="Sighting name"
+                value={title}
+                onChange={(e) => setTitle(e.target.value)}
+              />
+            </div>
+          </div>
+
+          <div className="field">
+            <label className="label">Address</label>
+            <input
+              id="address"
+              className="input is-danger"
+              type="text"
+              placeholder="Street address"
+              value={address}
+              onChange={(e) => setAddress(e.target.value)}
+            />
+          </div>
+
+          <div className="field is-grouped">
+            <p className="control is-expanded">
+              <label className="label">City</label>
+            </p>
+            <label className="label">State</label>
+          </div>
+
+          <div className="field is-grouped">
+            <input
+              className="input is-danger"
+              type="text"
+              placeholder="City"
+              value={city}
+              onChange={(e) => setCity(e.target.value)}
+            />
+
+            <div className="control">
+              <div className="select">
+                <select
+                  value={state}
+                  onChange={(e) => setState(e.target.value)}
+                >
+                  <option>Select dropdown</option>
+                  <option>AL</option>
+                  <option>AK</option>
+                  <option>AZ</option>
+                  <option>GA</option>
+                  <option>NC</option>
+                  <option>SC</option>
+                </select>
+              </div>
+            </div>
+          </div>
+
+          <div className="field">
+            <label className="label">Description</label>
+            <div className="control">
+              <textarea
+                className="textarea"
+                placeholder="Description"
+                value={description}
+                onChange={(e) => setDescription(e.target.value)}
+              ></textarea>
+            </div>
+          </div>
+
+          <div className="field is-grouped">
+            <div className="control">
+              <button
+                className="button is-link"
+                onClick={(e) => {
+                  handleFormSubmit(e, {
+                    username,
+                    title,
+                    address,
+                    city,
+                    state,
+                    description,
+                    imageUrl,
+                  });
+                }}
+              >
+                Submit
+              </button>
+            </div>
+            <div className="control">
+              <button className="button is-link is-light">Cancel</button>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default SightingsForm;
